Split maps overlay into separate parts
